requirements.txt
requests==2.31.0
pandas==2.2.0
numpy==1.26.4
python-dotenv==1.0.1
matplotlib==3.8.3
seaborn==0.13.2
scikit-learn==1.4.1.post1
scipy==1.12.0
tqdm==4.66.2
jupyter==1.0.0
